Moved writerecrod to its own class
