+++ VisioAutomation_2010/VisioAutomation/VisioAutomation.csproj
@@ -113,6 +113,7 @@
     <Compile Include="ShapeSheet\Writers\SidSrcWriter.cs" />
     <Compile Include="ShapeSheet\Writers\SrcWriter.cs" />
     <Compile Include="ShapeSheet\Streams\StreamType.cs" />
+    <Compile Include="ShapeSheet\Writers\WriteRecord.cs" />
     <Compile Include="ShapeSheet\Writers\WriteRecordList.cs" />
     <Compile Include="ShapeSheet\Streams\StreamArray.cs" />
     <Compile Include="Shapes\CustomPropertyType.cs" />
